refactor: move transform events to hotkeys.js
